Moved editor out of index file
